Made simulator optional.
Ideally we'd only require it on Windows, but .NET Core tries to preemptively load System.Windows.Forms even on Linux where we won't even load the simulator. Ideally it loads the simulator assembly on demand via an environment variable, but I had issues with it finding System.Windows.Forms in that case.